Fixed warnings on Windows, removed commented code and replaced tabulations with two spaces

+++ src/archive_module.xq.src/archive_module.h
@@ -199,7 +199,7 @@ namespace zorba { namespace archive {
       protected:
         String theEntryPath;
         String theEncoding;
-        int theSize;
+        long long theSize;
         time_t theLastModified;
         String theCompression;
         bool theSkipExtras;
@@ -211,7 +211,7 @@ namespace zorba { namespace archive {
 
         const String& getEncoding() const { return theEncoding; }
 
-        int getSize() const { return theSize; }
+        long long getSize() const { return theSize; }
 
         const time_t& getLastModified() const { return theLastModified; }
